refactor: move prisma provide to infrastructure layer
