+++ src/customers/application/repositories/customer.repository.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from '@nestjs/testing';
 
 import { CustomerRepository } from './customer.repository';
-import { PrismaProvider } from '../../../prisma/application/providers/prisma.provider';
+import { PrismaProvider } from '../../../prisma/infrastructure/providers/prisma.provider';
 import { prismaMock } from '../../domain/mocks/customers-providers.mock';
 import { mockCustomer1, mockCustomer2, mockCustomer3, mockCustomers } from '../../domain/mocks/customers.mock';
 import { CustomerEntity } from '../../domain/entities/customer.entity';
