+++ src/inventory/application/repositories/inventory.repository.ts
@@ -1,6 +1,6 @@
 import { Injectable } from '@nestjs/common';
 
-import { PrismaProvider } from '../../../prisma/application/providers/prisma.provider';
+import { PrismaProvider } from '../../../prisma/infrastructure/providers/prisma.provider';
 import { Pagination } from '../../../common/domain/types';
 import { InventoryEntity } from '../../domain/entities/inventory.entity';
 import { SelectedFields } from '../../domain/types';
