+++ src/prisma/prisma.module.ts
@@ -1,6 +1,6 @@
 import { Global, Module } from '@nestjs/common';
 
-import { PrismaProvider } from './application/providers/prisma.provider';
+import { PrismaProvider } from './infrastructure/providers/prisma.provider';
 
 @Global()
 @Module({
