+++ src/products/application/repositories/product.repository.spec.ts
@@ -1,7 +1,7 @@
 import { Test, TestingModule } from "@nestjs/testing";
 
 import { ProductRepository } from "./product.repository";
-import { PrismaProvider } from "../../../prisma/application/providers/prisma.provider";
+import { PrismaProvider } from "../../../prisma/infrastructure/providers/prisma.provider";
 import { prismaMock } from "../../domain/mocks/product-providers.mock";
 import { ProductEntity } from "../../domain/entities/product.entity";
 import { productMock1, productMock2, productMock3, productMock4, productMock5, productMock6 } from "../../domain/mocks/product.mock";
